+++ src/components/CategoriesContainer.jsx
@@ -1,11 +1,8 @@
 import { useDispatch, useSelector } from 'react-redux';
 
-import {
-  selectCategory,
-  loadRestaurants,
-} from './actions';
+import { selectCategory, loadRestaurants } from '../redux/actions';
 
-import { get } from './utils';
+import { get } from '../utils/utils';
 
 export default function CategoriesContainer() {
   const dispatch = useDispatch();
@@ -22,15 +19,10 @@ export default function CategoriesContainer() {
     <ul>
       {categories.map((category) => (
         <li key={category.id}>
-          <button
-            type="button"
-            onClick={() => handleClick(category.id)}
-          >
+          <button type="button" onClick={() => handleClick(category.id)}>
             {category.name}
             {selectedCategory ? (
-              <>
-                {category.id === selectedCategory.id ? '(V)' : null}
-              </>
+              <>{category.id === selectedCategory.id ? '(V)' : null}</>
             ) : null}
           </button>
         </li>
