+++ src/components/RegionsContainer.jsx
@@ -1,11 +1,8 @@
 import { useDispatch, useSelector } from 'react-redux';
 
-import {
-  selectRegion,
-  loadRestaurants,
-} from './actions';
+import { selectRegion, loadRestaurants } from '../redux/actions';
 
-import { get } from './utils';
+import { get } from '../utils/utils';
 
 export default function RegionsContainer() {
   const dispatch = useDispatch();
@@ -22,15 +19,10 @@ export default function RegionsContainer() {
     <ul>
       {regions.map((region) => (
         <li key={region.id}>
-          <button
-            type="button"
-            onClick={() => handleClick(region.id)}
-          >
+          <button type="button" onClick={() => handleClick(region.id)}>
             {region.name}
             {selectedRegion ? (
-              <>
-                {region.id === selectedRegion.id ? '(V)' : null}
-              </>
+              <>{region.id === selectedRegion.id ? '(V)' : null}</>
             ) : null}
           </button>
         </li>
