+++ src/index.jsx
@@ -1,22 +1,18 @@
 import ReactDOM from 'react-dom';
 
-import {
-  BrowserRouter,
-} from 'react-router-dom';
+import { BrowserRouter } from 'react-router-dom';
 
 import { Provider } from 'react-redux';
 
 import App from './App';
 
-import store from './store';
+import store from './redux/store';
 
 ReactDOM.render(
-  (
-    <Provider store={store}>
-      <BrowserRouter>
-        <App />
-      </BrowserRouter>
-    </Provider>
-  ),
-  document.getElementById('app'),
+  <Provider store={store}>
+    <BrowserRouter>
+      <App />
+    </BrowserRouter>
+  </Provider>,
+  document.getElementById('app')
 );
