+++ src/pages/RestaurantPage.jsx
@@ -1,11 +1,9 @@
 import { useParams } from 'react-router-dom';
 
-import RestaurantContainer from './RestaurantContainer';
+import RestaurantContainer from '../components/RestaurantContainer';
 
 export default function RestaurantPage({ params }) {
   const { id } = params || useParams();
 
-  return (
-    <RestaurantContainer restaurantId={id} />
-  );
+  return <RestaurantContainer restaurantId={id} />;
 }
